requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymongo==4.6.0
pinecone-client>=3.0.0
pydantic==2.4.2
google-generativeai==0.3.1
python-multipart==0.0.6
pytz==2023.3
httpx==0.25.1
uuid==1.30
psutil==5.9.5
prometheus-client==0.17.1
pytest==7.4.0
pytest-cov==4.1.0
starlette==0.27.0 